Tested on device!

+++ smart_home/src/acs712.h
@@ -20,7 +20,7 @@ private:
   uint16_t _defaultNull;
   uint8_t _sensorNumber;
   uint64_t _samplingSum;
-  uint16_t _samplingCount;
+  uint32_t _samplingCount;
 };
 
 #endif
